feat: RV32IY CHERI — purecap capability CPU (#2)
* feat: RV32IY CHERI Phase 1 — 65-bit capability regfile + PCC pipeline

Upgrade cpu_core.tbn from RV32I to capability-aware RV32IY pipeline:
- 32×65-bit register file (Cap = tag + metadata + address)
- PCC-based fetch via next_pcc queue with root capability init
- Decode extracts 32-bit addresses from capability registers
- Execute wraps integer results as null capabilities
- cheri_pkg.sv with CHERI Concentrate field extraction, bounds
  encode/decode, permission operations (pure SV, synthesisable)
- Toolchain auto-fetch script with SHA-256 verification

All 38 rv32ui tests pass. No bounds checking yet (Phase 2).


* feat: RV32IY CHERI Phase 2 — capability instruction decode + execute

Add CHERI instruction support to the Execute stage:
- YADD (cincoffset): increment capability address by integer offset
- YADDRW (csetaddr): set capability address field
- YBNDSW (csetbounds): narrow capability bounds
- YPERMC (candperm): restrict permissions monotonically
- YTAGR/YPERMR/YTYPER/YTOPR/YBASER/YLENR: capability field inspection
- PACKY (ccleartag): clear tag bit
- AUIPC now produces PCC-derived capability (not integer)
- JAL/JALR link registers now carry PCC-derived capabilities
- JALR jump target PCC derived from rs1 capability

Decoded record extended with rs1_cap, rs2_cap, funct7, rs2_idx for
full capability operand passing. Execute dispatches via
cheri_compute_result() and cheri_derive_next_pcc() in cheri_pkg.sv.

All 38 rv32ui tests pass. New cheri_basic test exercises 7 CHERI ops.


* chore: configure CHERI toolchain download URL

Set default URL for toolchain_coralnpu_v2_cheri.tar.xz from
GCS public artifacts bucket. SHA-256 checksum will be recorded
after first successful download.


* fix: use RVY encoding constants, pin toolchain SHA-256

Revert to RVY spec encoding (standard OP opcode 0x33 with new funct7
values) rather than CTSRD legacy encoding (custom-2 opcode 0x5B).
The CTSRD toolchain uses legacy encoding but no RVY toolchain exists
yet — we hand-assemble RVY tests for now.

Also pin the CHERI toolchain SHA-256 checksum for reproducible fetches.


* feat: widen integer literals from u64 to u128, enable 65-bit CHERI init

The Tourbillon compiler stored integer literals as u64, which truncated
the tag bit (bit 64) of 65-bit capability init values. This forced a
testbench hack to patch the PCC tag after reset.

Widen Literal::Int, Token::Int, init_tokens, and Cell init to u128
throughout the compiler pipeline (lexer → parser → AST → IR → lower).
The next_pcc FIFO now initializes with the full root PCC including tag=1:
  init = 0x1FFC0818080000000

Also adds bounds checking to Execute (Phase 3a): loads/stores through
tagged capabilities are checked against bounds and permissions.
Violations suppress the memory access and zero the destination (soft trap).
Tag=0 caps skip checks for backward compatibility with rv32ui tests.


* feat: RV32IY CHERI Phase 3 — bounds checking + CHERI Concentrate decode

Fix CHERI Concentrate bounds decode to reconstruct full 32-bit base/top
from compressed mantissa + address upper bits (with carry correction).
Previous implementation only used the mantissa, producing wrong results.

Bounds checking now works end-to-end:
- AUIPC derives tag=1 capability from PCC (root cap)
- YBNDSW narrows bounds to [addr, addr+length)
- Loads/stores through bounded caps check [base, top) and permissions
- Out-of-bounds access → soft trap: suppress memory op, zero destination
- Tag=0 caps (integer pointers) skip check for rv32ui compatibility

Test: cheri_bounds.S exercises in-bounds loads, out-of-bounds loads
(offset > bounds), and boundary-edge cases. Both soft-trap (zero result)
and pass-through cases verified.

38/38 rv32ui + 7 CHERI basic + 6 bounds tests all pass.


* feat: LC/SC capability load/store, tag memory, tuple width-cast fix

Complete Phase 3 leftovers for CHERI memory safety:

LC (load capability): 2-beat transfer on 32-bit bus — reads low word,
then high word + tag in parallel, assembles 65-bit capability result.
SC (store capability): writes low word, then high word + tag bit.
Tag memory: separate Memory(Addr → Bits 1, depth=2048) alongside data
memory, one tag per 8-byte capability-aligned region.
Integer stores (SW/SH/SB) clear tag at the affected 8-byte region.

Compiler fix: tuple elements in SV concatenation now get per-element
width casts from the queue's element type. Previously, Literal::Int(0)
emitted as bare `0` (32-bit default in SV) inside `{addr, tag}` tuples,
causing width mismatch. Also fixed width_cast to emit `1'(expr)` for
single-bit values instead of returning bare expression.

New ExecStash fields: rs2_cap (for SC data), stash_lo (for LC beat 1).
Execute phases: 3-4 for LC (low word → high word+tag), 5 for SC.

38/38 rv32ui + 7 CHERI basic + 6 bounds + 4 capmem tests all pass.


* feat: RV32IY CHERI Phase 4 — Marie SoC integration + async FIFO reset fix

Integrate CHERI capabilities into the Marie Antoinette SoC:
- Tag memory (Memory(Addr → Bits 1, depth=8192)) in CPU domain
- CPUCore tmem ports wired through Marie to soc_top.sv
- Stub tag memory in soc_top (always returns tag=0, sufficient for
  integer-only rv32ui tests; real tag storage for Phase 5)

Fix async FIFO reset wiring (pre-existing bug): wr_rst_n and rd_rst_n
were hardcoded to bare `rst_n` (undefined in multi-domain modules).
Now uses domain-specific resets (e.g., cpu_rst_n, xbar_rst_n). This
was causing the SoC to deadlock — async FIFOs stayed in reset forever.

Fix soc_tb.cpp cycle-limit loop: `t < (max_cycles * 10 + t)` was always
true (t on both sides). Now uses fixed endpoint `t < t_end`.

38/38 rv32ui standalone + 38/38 rv32ui SoC + 17 CHERI + 213 cargo = all green.


* feat: RV32IY CHERI Phase 5+6 — full test coverage, SoC tag memory, FPGA RTL

Phase 5 (Test Completion):
- Add cheri_inspect.S: YTOPR/YBASER/YLENR/YTYPER/YAMASK round-trip tests
- Add cheri_perms.S: YPERMC restriction + load/store enforcement tests
- Add cheri_bounds_rt.S: CHERI Concentrate encode/decode at 5 scales
- All CHERI tests now run through SoC path (soc-cheri-test target)
- Fix CHERI Concentrate base decode: mask E[2:0] from BE in EF=0 path

Phase 6 (FPGA):
- Add fpga_tag_mem.sv: synthesisable 1-bit distributed RAM tag memory
- Wire tag memory in marie_top.sv (STANDALONE + external port modes)
- Add cheri_pkg.sv and fpga_tag_mem.sv to RTL export packages
- Replace stub tag memory in soc_top.sv with real mem_model_reg

Verification: 38/38 rv32ui + 6/6 CHERI (standalone + SoC), rtl-check clean.


* ci: add GitHub Actions workflow for Rust + Verilator tests

Two jobs:
- rust: fmt, clippy, cargo test
- sim: rv32ui (38) + CHERI (6) standalone and SoC, FPGA RTL lint


* fix(ci): list SV packages before importing files for older Verilator

Verilator 5.020 (Ubuntu apt) requires packages to appear before files
that import them on the command line. Newer versions resolve this
automatically but the stricter ordering is correct regardless.


* perf: drain SoC UART TX by idle detection instead of fixed 100k cycles

Add uart_tx_idle output from soc_top.sv TX deserializer. Testbench now
exits drain when TX has been idle for 2 byte periods (680 dev_clk ticks)
instead of blindly waiting 100k ticks. Non-UART tests drop from ~40k to
~400 cpu cycles; Hello World from ~40k to ~10k.


---------

+++ src/parse/parser.rs
@@ -28,7 +28,7 @@ where
                 just(Token::Comma)
                     .ignore_then(just(Token::Depth))
                     .ignore_then(just(Token::Assign))
-                    .ignore_then(select! { Token::Int(n) => n })
+                    .ignore_then(select! { Token::Int(n) => n as u64 })
                     .or_not(),
             )
             .then_ignore(just(Token::RParen))
@@ -44,7 +44,7 @@ where
                 just(Token::Comma)
                     .ignore_then(just(Token::Depth))
                     .ignore_then(just(Token::Assign))
-                    .ignore_then(select! { Token::Int(n) => n })
+                    .ignore_then(select! { Token::Int(n) => n as u64 })
                     .or_not(),
             )
             .then_ignore(just(Token::RParen))
@@ -89,7 +89,7 @@ where
             });
 
         let named_with_int = select! { Token::Ident(s) => s }
-            .then(select! { Token::Int(n) => n }.map_with(|n, e| {
+            .then(select! { Token::Int(n) => n as u64 }.map_with(|n, e| {
                 spn(
                     TypeExpr::Named {
                         name: format!("{n}"),
@@ -109,7 +109,7 @@ where
         });
 
         // Integer literals as type expressions (e.g. 32 in Array(32, Word))
-        let int_as_type = select! { Token::Int(n) => n }.map(|n| TypeExpr::Named {
+        let int_as_type = select! { Token::Int(n) => n as u64 }.map(|n| TypeExpr::Named {
             name: format!("{n}"),
             args: vec![],
         });
@@ -347,9 +347,9 @@ where
 
         // Bit slice: [int : int]
         let bit_slice_postfix = just(Token::LBrack)
-            .ignore_then(select! { Token::Int(hi) => hi })
+            .ignore_then(select! { Token::Int(hi) => hi as u64 })
             .then_ignore(just(Token::Colon))
-            .then(select! { Token::Int(lo) => lo })
+            .then(select! { Token::Int(lo) => lo as u64 })
             .then_ignore(just(Token::RBrack));
 
         let index_postfix = expr
@@ -632,7 +632,7 @@ where
         .then_ignore(just(Token::Colon))
         .then(select! { Token::Ident(s) => s }.map_with(|s: &str, e| spn(s.to_string(), e.span())))
         .then(
-            select! { Token::Int(n) => n }
+            select! { Token::Int(n) => n as u64 }
                 .delimited_by(just(Token::LBrack), just(Token::RBrack))
                 .or_not(),
         )
@@ -729,7 +729,7 @@ where
             just(Token::Comma)
                 .ignore_then(just(Token::Depth))
                 .ignore_then(just(Token::Assign))
-                .ignore_then(select! { Token::Int(n) => n })
+                .ignore_then(select! { Token::Int(n) => n as u64 })
                 .or_not(),
         )
         .then(
@@ -761,13 +761,13 @@ where
             just(Token::Comma)
                 .ignore_then(just(Token::Depth))
                 .ignore_then(just(Token::Assign))
-                .ignore_then(select! { Token::Int(n) => n }),
+                .ignore_then(select! { Token::Int(n) => n as u64 }),
         )
         .then(
             just(Token::Comma)
                 .ignore_then(just(Token::Latency))
                 .ignore_then(just(Token::Assign))
-                .ignore_then(select! { Token::Int(n) => n }),
+                .ignore_then(select! { Token::Int(n) => n as u64 }),
         )
         .then_ignore(just(Token::RParen))
         .map(|((((name, key_ty), val_ty), depth), latency)| MemoryDecl {
@@ -789,7 +789,7 @@ where
             just(Token::Comma)
                 .ignore_then(just(Token::Depth))
                 .ignore_then(just(Token::Assign))
-                .ignore_then(select! { Token::Int(n) => n })
+                .ignore_then(select! { Token::Int(n) => n as u64 })
                 .or_not(),
         )
         .then_ignore(just(Token::RParen))
@@ -974,7 +974,7 @@ where
     let const_def = just(Token::Const)
         .ignore_then(ident_sp)
         .then_ignore(just(Token::Assign))
-        .then(select! { Token::Int(n) => n })
+        .then(select! { Token::Int(n) => n as u64 })
         .map(|(name, value)| Item::Const(ConstDef { name, value }));
 
     // external fn name(params) [-> RetTy]
